requirements.txt
Django
djangorestframework
django-ckeditor
redis
googletrans
django-redis

